requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.5.0
tqdm>=4.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
sec-edgar-downloader>=5.0.0
streamlit>=1.30.0
numpy>=1.24.0